requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pdfplumber==0.11.4
langchain==0.3.0
langchain-text-splitters==0.3.0
sentence-transformers==3.1.1
faiss-cpu==1.8.0
groq==0.11.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.9.2
gradio==4.44.0
numpy==1.26.4
